requirements.txt
# DevOp packages
# pylint == 3.0.2
ruff == 0.6.6
pytest == 8.3.2
black == 24.3.0
pytest-cov == 4.0.0
nbval == 0.11.0
requests==2.26.0

# template specific packages
pandas == 2.0.3
matplotlib == 3.7.5
jupyter == 1.0.0

# databricks packages
databricks-sql-connector == 3.5.0
python-dotenv == 1.0.1